Refactor snaphot_brain_montage()
