+++ mne/viz/backends/_pysurfer_mayavi.py
@@ -3,8 +3,18 @@ Core visualization operations.
 """
 
 import warnings
+import numpy as np
 from ...surface import _normalize_vectors
-from ...utils import _import_mlab
+from ...utils import _import_mlab, _validate_type
+
+
+class Projection:
+    def __init__(self, xy=None, pts=None):
+        self.xy = xy
+        self.pts = pts
+
+    def visible(self, state):
+        self.pts.visible = state
 
 
 class Renderer:
@@ -93,6 +103,13 @@ class Renderer:
     def screenshot(self):
         return self.mlab.screenshot(self.fig)
 
+    def project(self, xyz, ch_names):
+        xy = _3d_to_2d(self.fig, xyz)
+        xy = dict(zip(ch_names, xy))
+        pts = self.fig.children[-1]
+
+        return Projection(xy=xy, pts=pts)
+
 
 def _mlab_figure(**kwargs):
     """Create a Mayavi figure using our defaults."""
@@ -130,3 +147,74 @@ def _create_mesh_surf(surf, fig=None, scalars=None, vtk_normals=True):
         mesh.data.point_data.normals = nn
         mesh.data.cell_data.normals = None
     return mesh
+
+
+def _3d_to_2d(fig, xyz):
+    """Convert 3d points to a 2d perspective using a Mayavi Scene."""
+    from mayavi.core.scene import Scene
+
+    _validate_type(fig, Scene, "fig", "Scene")
+    xyz = np.column_stack([xyz, np.ones(xyz.shape[0])])
+
+    # Transform points into 'unnormalized' view coordinates
+    comb_trans_mat = _get_world_to_view_matrix(fig.scene)
+    view_coords = np.dot(comb_trans_mat, xyz.T).T
+
+    # Divide through by the fourth element for normalized view coords
+    norm_view_coords = view_coords / (view_coords[:, 3].reshape(-1, 1))
+
+    # Transform from normalized view coordinates to display coordinates.
+    view_to_disp_mat = _get_view_to_display_matrix(fig.scene)
+    xy = np.dot(view_to_disp_mat, norm_view_coords.T).T
+
+    # Pull the first two columns since they're meaningful for 2d plotting
+    xy = xy[:, :2]
+    return xy
+
+
+def _get_world_to_view_matrix(scene):
+    """Return the 4x4 matrix to transform xyz space to the current view.
+
+    This is a concatenation of the model view and perspective transforms.
+    """
+    from mayavi.core.ui.mayavi_scene import MayaviScene
+    from tvtk.pyface.tvtk_scene import TVTKScene
+
+    _validate_type(scene, (MayaviScene, TVTKScene), "scene",
+                   "TVTKScene/MayaviScene")
+    cam = scene.camera
+
+    # The VTK method needs the aspect ratio and near and far
+    # clipping planes in order to return the proper transform.
+    scene_size = tuple(scene.get_size())
+    clip_range = cam.clipping_range
+    aspect_ratio = float(scene_size[0]) / scene_size[1]
+
+    # Get the vtk matrix object using the aspect ratio we defined
+    vtk_comb_trans_mat = cam.get_composite_projection_transform_matrix(
+        aspect_ratio, clip_range[0], clip_range[1])
+    vtk_comb_trans_mat = vtk_comb_trans_mat.to_array()
+    return vtk_comb_trans_mat
+
+
+def _get_view_to_display_matrix(scene):
+    """Return the 4x4 matrix to convert view coordinates to display coordinates.
+
+    It's assumed that the view should take up the entire window and that the
+    origin of the window is in the upper left corner.
+    """  # noqa: E501
+    from mayavi.core.ui.mayavi_scene import MayaviScene
+    from tvtk.pyface.tvtk_scene import TVTKScene
+
+    _validate_type(scene, (MayaviScene, TVTKScene), "scene",
+                   "TVTKScene/MayaviScene")
+
+    # normalized view coordinates have the origin in the middle of the space
+    # so we need to scale by width and height of the display window and shift
+    # by half width and half height. The matrix accomplishes that.
+    x, y = tuple(scene.get_size())
+    view_to_disp_mat = np.array([[x / 2.0,       0.,   0.,   x / 2.0],
+                                 [0.,      -y / 2.0,   0.,   y / 2.0],
+                                 [0.,            0.,   1.,        0.],
+                                 [0.,            0.,   0.,        1.]])
+    return view_to_disp_mat
